+++ src/main/java/com/bluetoya/taradiddle/common/filter/JwtAuthenticationFilter.java
@@ -1,15 +1,14 @@
 package com.bluetoya.taradiddle.common.filter;
 
 import com.bluetoya.taradiddle.common.config.JwtProvider;
-import com.bluetoya.taradiddle.feature.auth.AuthUser;
-import com.bluetoya.taradiddle.feature.auth.AuthUserService;
+import com.bluetoya.taradiddle.feature.auth.entity.AuthUser;
+import com.bluetoya.taradiddle.feature.auth.service.AuthUserService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
