--- src/main/java/com/bluetoya/taradiddle/feature/auth/AuthService.java
@@ -1,8 +1,0 @@
-package com.bluetoya.taradiddle.feature.auth;
-
-public interface AuthService {
-
-    LoginResponse login(String userId, String password);
-
-    SignInResponse signIn(SignInRequest request);
-}
