--- src/main/java/com/bluetoya/taradiddle/feature/auth/LoginResponse.java
@@ -1,5 +1,0 @@
-package com.bluetoya.taradiddle.feature.auth;
-
-public record LoginResponse(String token) {
-
-}
