+++ src/main/java/com/bluetoya/taradiddle/feature/auth/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.bluetoya.taradiddle.feature.auth;
+package com.bluetoya.taradiddle.feature.auth.dto;
 
 public record LoginRequest(String userId, String password) {
