+++ src/main/java/com/bluetoya/taradiddle/feature/auth/dto/LoginResponse.java
@@ -1,0 +1,5 @@
+package com.bluetoya.taradiddle.feature.auth.dto;
+
+public record LoginResponse(String token) {
+
+}
