+++ src/main/java/com/bluetoya/taradiddle/feature/auth/dto/SignInRequest.java
@@ -1,4 +1,4 @@
-package com.bluetoya.taradiddle.feature.auth;
+package com.bluetoya.taradiddle.feature.auth.dto;
 
 public record SignInRequest(String userId, String password, String confirmPassword, String username, String firstName, String lastName) {
