+++ src/main/java/com/bluetoya/taradiddle/feature/auth/dto/SignInResponse.java
@@ -1,5 +1,6 @@
-package com.bluetoya.taradiddle.feature.auth;
+package com.bluetoya.taradiddle.feature.auth.dto;
 
+import com.bluetoya.taradiddle.feature.auth.entity.AuthUser;
 import com.bluetoya.taradiddle.feature.user.User;
 
 public record SignInResponse(AuthUser authUser, User user) {
