+++ src/main/java/com/bluetoya/taradiddle/feature/auth/repository/AuthRepository.java
@@ -1,5 +1,6 @@
-package com.bluetoya.taradiddle.feature.auth;
+package com.bluetoya.taradiddle.feature.auth.repository;
 
+import com.bluetoya.taradiddle.feature.auth.entity.AuthUser;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface AuthRepository extends MongoRepository<AuthUser, String> {
