+++ src/main/java/com/bluetoya/taradiddle/feature/auth/validator/Validator.java
@@ -1,4 +1,4 @@
-package com.bluetoya.taradiddle.feature.auth;
+package com.bluetoya.taradiddle.feature.auth.validator;
 
 public interface Validator<T> {
